한 줄 소개도 따로 상호작용시 변경되지 않게 수정
